Use wxjpeg_boolean in libtiff sources.
This fixes compilation problem when using libtiff with the built-in modified
version of libjpeg that uses wxjpeg_boolean instead of boolean as it doesn't
define boolean at all in its headers, but just wxjpeg_boolean.

See #15179.

(this is a "cherry-pick" of 94474c2e0b2b13c2e0adefd9c307eb39433283b1
from the main repository)