Add config files for builds not using configure
Check in files used on systems without configure to bring the submodule
in sync with the way wxWidgets has been building libtiff so far.

It really ought to change because this is very messy and there is, in
particular, some horrible confusion between tif_config.h and tiffconf.h
for Xcode builds.

But for now just try to keep things working as before.

Note that this change requires always using angle brackets instead of
double quotes when including tiffconf.h and tif_config.h from libtiff
own code, as otherwise these pre-generated versions from the same
directory as the file including the header would be used even under Unix
instead of the ones generated by configure.